✨ cleanup/refactor api code structure 🔮
