+++ api/src/items/entity.py
@@ -1,0 +1,40 @@
+import uuid
+from datetime import datetime
+from sqlmodel import (
+    Field,
+    SQLModel,
+    UniqueConstraint,
+    JSON,
+    Relationship,
+)
+from sqlalchemy import Column, DateTime, func
+from typing import Dict, Optional
+
+from ..namespaces.entity import NamespaceEntity
+
+
+class ItemEntity(SQLModel, table=True):
+    __tablename__ = "item"
+    __table_args__ = (
+        UniqueConstraint(
+            "namespace_name", "name", name="item_name_is_unique_in_namespace"
+        ),
+    )
+
+    uid: uuid.UUID = Field(primary_key=True, default_factory=uuid.uuid4)
+    namespace_name: str = Field(foreign_key="namespace.name")
+    namespace: NamespaceEntity = Relationship()
+    name: str = Field()
+    creationTimestamp: Optional[datetime] = Field(
+        sa_column=Column(
+            DateTime(timezone=True), server_default=func.now(), nullable=True
+        ),
+    )
+    updateTimestamp: Optional[datetime] = Field(
+        sa_column=Column(DateTime(timezone=True), onupdate=func.now(), nullable=True),
+    )
+    deletedTimestamp: Optional[datetime] = Field(
+        sa_column=Column(DateTime(timezone=True), nullable=True),
+    )
+    labels: Dict[str, str] = Field(default={}, sa_type=JSON)
+    annotations: Dict[str, str] = Field(default={}, sa_type=JSON)
